Updated to Tytung's GPS 1.3

+++ gps/leo-gps.c
@@ -827,6 +827,7 @@ static int gps_xtra_inject_xtra_data(char* data, int length) {
     GpsState*  s = _gps_state;
     if (!s->init)
         return 0;
+
     int rpc_ret_val = -1;
     int ret_val = -1;
     unsigned char *xtra_data_ptr;
@@ -983,7 +984,7 @@ static int gps_stop() {
 
 static int gps_inject_time(GpsUtcTime time, int64_t timeReference, int uncertainty) {
     D("%s() is called", __FUNCTION__);
-    D("time=%d, timeReference=%d, uncertainty=%d", (int) time, (int) timeReference, uncertainty);
+    D("time=%lld, timeReference=%lld, uncertainty=%d", time, timeReference, uncertainty);
     GpsState*  s = _gps_state;
     if (!s->init)
         return 0;
